# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Clone API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SessionInput {
  /**
     * @minLength 1
     * @maxLength 32
     */
  name: string;
}

export interface SessionUser {
  id: string;
  name: string;
  color: string;
  createdAt: string;
}

export interface User {
  id: string;
  name: string;
  color: string;
  lastSeen: string;
  createdAt: string;
}

export interface ChatInput {
  /** @nullable */
  name?: string | null;
  isGroup?: boolean;
  participantIds: string[];
}

export interface Chat {
  id: string;
  name: string;
  isGroup: boolean;
  participants: User[];
  createdAt: string;
}

export interface Message {
  id: string;
  chatId: string;
  senderId: string;
  senderName: string;
  senderColor: string;
  content: string;
  readBy: string[];
  createdAt: string;
}

export interface ChatSummary {
  id: string;
  name: string;
  isGroup: boolean;
  participants: User[];
  lastMessage?: Message;
  unreadCount: number;
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
  senderId: string;
}

export interface ReadResult {
  updated: number;
}

export interface AppStats {
  totalMessages: number;
  totalUsers: number;
  totalChats: number;
}