# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: WhatsApp Clone API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: session
    description: User session management
  - name: users
    description: User operations
  - name: chats
    description: Chat operations
  - name: messages
    description: Message operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /session:
    post:
      operationId: createSession
      tags: [session]
      summary: Create or resume a user session by name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionInput"
      responses:
        "200":
          description: Session created or resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
    get:
      operationId: getSession
      tags: [session]
      summary: Get current session user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
        "401":
          description: Not authenticated

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all registered users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found

  /chats:
    get:
      operationId: listChats
      tags: [chats]
      summary: List all chats for current user (with last message + unread count)
      responses:
        "200":
          description: Chat list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatSummary"
    post:
      operationId: createChat
      tags: [chats]
      summary: Create a new chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "201":
          description: Chat created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"

  /chats/{chatId}:
    get:
      operationId: getChat
      tags: [chats]
      summary: Get chat details
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
        "404":
          description: Not found

  /chats/{chatId}/messages:
    get:
      operationId: getChatMessages
      tags: [messages]
      summary: Get messages in a chat (newest last)
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message to a chat
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /chats/{chatId}/read:
    post:
      operationId: markChatRead
      tags: [messages]
      summary: Mark all messages in a chat as read
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadResult"

  /stats:
    get:
      operationId: getStats
      tags: [chats]
      summary: App-wide stats (total messages, active users, chats)
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppStats"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    SessionInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 32

    SessionUser:
      type: object
      required: [id, name, color, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
        createdAt:
          type: string

    User:
      type: object
      required: [id, name, color, lastSeen, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
        lastSeen:
          type: string
        createdAt:
          type: string

    ChatInput:
      type: object
      required: [participantIds]
      properties:
        name:
          type: ["string", "null"]
        isGroup:
          type: boolean
        participantIds:
          type: array
          items:
            type: string

    Chat:
      type: object
      required: [id, name, isGroup, participants, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        isGroup:
          type: boolean
        participants:
          type: array
          items:
            $ref: "#/components/schemas/User"
        createdAt:
          type: string

    ChatSummary:
      type: object
      required: [id, name, isGroup, participants, unreadCount, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        isGroup:
          type: boolean
        participants:
          type: array
          items:
            $ref: "#/components/schemas/User"
        lastMessage:
          $ref: "#/components/schemas/Message"
        unreadCount:
          type: integer
        createdAt:
          type: string

    Message:
      type: object
      required: [id, chatId, senderId, senderName, senderColor, content, readBy, createdAt]
      properties:
        id:
          type: string
        chatId:
          type: string
        senderId:
          type: string
        senderName:
          type: string
        senderColor:
          type: string
        content:
          type: string
        readBy:
          type: array
          items:
            type: string
        createdAt:
          type: string

    MessageInput:
      type: object
      required: [content, senderId]
      properties:
        content:
          type: string
          minLength: 1
        senderId:
          type: string

    ReadResult:
      type: object
      required: [updated]
      properties:
        updated:
          type: integer

    AppStats:
      type: object
      required: [totalMessages, totalUsers, totalChats]
      properties:
        totalMessages:
          type: integer
        totalUsers:
          type: integer
        totalChats:
          type: integer