# lib/api-zod/src/generated/types/appStats.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Clone API
 * OpenAPI spec version: 0.1.0
 */

export interface AppStats {
  totalMessages: number;
  totalUsers: number;
  totalChats: number;
}