# lib/api-zod/src/generated/types/chat.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Clone API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface Chat {
  id: string;
  name: string;
  isGroup: boolean;
  participants: User[];
  createdAt: string;
}