# lib/api-zod/src/generated/types/chatInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Clone API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatInput {
  /** @nullable */
  name?: string | null;
  isGroup?: boolean;
  participantIds: string[];
}