# lib/api-zod/src/generated/types/chatSummary.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Clone API
 * OpenAPI spec version: 0.1.0
 */
import type { Message } from './message';
import type { User } from './user';

export interface ChatSummary {
  id: string;
  name: string;
  isGroup: boolean;
  participants: User[];
  lastMessage?: Message;
  unreadCount: number;
  createdAt: string;
}