# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Clone API
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: string;
  chatId: string;
  senderId: string;
  senderName: string;
  senderColor: string;
  content: string;
  readBy: string[];
  createdAt: string;
}