# lib/api-zod/src/generated/types/sessionInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Clone API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionInput {
  /**
     * @minLength 1
     * @maxLength 32
     */
  name: string;
}