# lib/api-zod/src/generated/types/sessionUser.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Clone API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionUser {
  id: string;
  name: string;
  color: string;
  createdAt: string;
}